# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Tracker API",
        "description": "Документация для API трекинга задач",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Список задач",
                "operationId": "49e73b9da657e3e2dac88d10e4ffaa4b",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "new",
                                "in_progress",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Фильтр по приоритету",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "high",
                                "normal",
                                "low"
                            ]
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Фильтр по пользователю",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Новая задача"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Описание задачи"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "new",
                                                    "in_progress",
                                                    "completed",
                                                    "cancelled"
                                                ],
                                                "example": "new"
                                            },
                                            "priority": {
                                                "type": "string",
                                                "enum": [
                                                    "high",
                                                    "normal",
                                                    "low"
                                                ],
                                                "example": "high"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-17T12:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-17T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать новую задачу",
                "operationId": "c55168b216f825bfb38201acb8ddba12",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Новая задача"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание задачи"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "high",
                                            "normal",
                                            "low"
                                        ],
                                        "example": "high"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Новая задача"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Описание задачи"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "new",
                                                "in_progress",
                                                "completed",
                                                "cancelled"
                                            ],
                                            "example": "new"
                                        },
                                        "priority": {
                                            "type": "string",
                                            "enum": [
                                                "high",
                                                "normal",
                                                "low"
                                            ],
                                            "example": "high"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-09-17T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-09-17T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{id}/status": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить статус задачи",
                "operationId": "02ebd1b62bf5e3c0856245a3b19422b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "in_progress",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "completed"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Новая задача"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Описание задачи"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "new",
                                                "in_progress",
                                                "completed",
                                                "cancelled"
                                            ],
                                            "example": "completed"
                                        },
                                        "priority": {
                                            "type": "string",
                                            "enum": [
                                                "high",
                                                "normal",
                                                "low"
                                            ],
                                            "example": "high"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-09-17T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-09-17T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{id}/comments": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Добавить комментарий к задаче",
                "operationId": "695e580e228b51716d8bdeab0258a30d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment",
                                    "user_id"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Комментарий к задаче"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комментарий добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "task_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "comment": {
                                            "type": "string",
                                            "example": "Комментарий к задаче"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-09-17T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}